recording audio workaround

+++ app/src/main/java/com/example/company/birds_classifier_android/Classifier.java
@@ -88,7 +88,7 @@ public interface Classifier {
         }
     }
 
-    List<Recognition> recognizeImage(double[] pixels);
+    List<Recognition> recognizeImage(float[] pixels);
 
     void enableStatLogging(final boolean debug);
